cmnlib-term: fixed type use wrt u8strings -- uint16_t for printf, u8 string litterals

+++ cmnlib-term/src-tests/vt-input/TestRunner--VtInputFromCharacters.cpp
@@ -27,16 +27,16 @@ Test(VtInputFromCharacters, should_return_an_unknown_virtual_terminal_input_on_r
     for (char8_t c : toBeTested) {
         cmspk::term::VtInputFromCharacters dut;
         dut.reset();
-        cr_assert(dut.append(c), "Failed for input value %d", (uint8_t)c);
-        cr_assert(not(dut.canAppend()), "Failed for input value %d", (uint8_t)c);
-        cr_assert(dut.canGetData(), "Failed for input value %d", (uint8_t)c);
+        cr_assert(dut.append(c), "Failed for input value %d", (uint16_t)c);
+        cr_assert(not(dut.canAppend()), "Failed for input value %d", (uint16_t)c);
+        cr_assert(dut.canGetData(), "Failed for input value %d", (uint16_t)c);
         std::optional<cmspk::term::VtInput> vtin = dut.getData();
-        cr_assert(vtin, "Failed for input value %d", (uint8_t)c);
-        cr_assert(std::holds_alternative<cmspk::term::VtInputUnknown>(*vtin), "Failed for input value %d", (uint8_t)c);
+        cr_assert(vtin, "Failed for input value %d", (uint16_t)c);
+        cr_assert(std::holds_alternative<cmspk::term::VtInputUnknown>(*vtin), "Failed for input value %d", (uint16_t)c);
         cmspk::term::VtInputUnknown vtu = std::get<cmspk::term::VtInputUnknown>(*vtin);
         char8_t raw = vtu.rawValue;
-        cr_assert((c == raw), "Failed for input value %d", (uint8_t)c);
-        cr_assert(not(dut.canGetData()), "Failed for input value %d", (uint8_t)c);
+        cr_assert((c == raw), "Failed for input value %d", (uint16_t)c);
+        cr_assert(not(dut.canGetData()), "Failed for input value %d", (uint16_t)c);
     }
 }
 Test(VtInputFromCharacters, for_now_should_return_an_unknown_virtual_terminal_input_on_reading_any_octet) {
@@ -44,16 +44,16 @@ Test(VtInputFromCharacters, for_now_should_return_an_unknown_virtual_terminal_in
         char8_t c = (char8_t)cint;
         cmspk::term::VtInputFromCharacters dut;
         dut.reset();
-        cr_assert(dut.append(c), "Failed for input value %d", (uint8_t)c);
-        cr_assert(not(dut.canAppend()), "Failed for input value %d", (uint8_t)c);
-        cr_assert(dut.canGetData(), "Failed for input value %d", (uint8_t)c);
+        cr_assert(dut.append(c), "Failed for input value %d", (uint16_t)c);
+        cr_assert(not(dut.canAppend()), "Failed for input value %d", (uint16_t)c);
+        cr_assert(dut.canGetData(), "Failed for input value %d", (uint16_t)c);
         std::optional<cmspk::term::VtInput> vtin = dut.getData();
-        cr_assert(vtin, "Failed for input value %d", (uint8_t)c);
-        cr_assert(std::holds_alternative<cmspk::term::VtInputUnknown>(*vtin), "Failed for input value %d", (uint8_t)c);
+        cr_assert(vtin, "Failed for input value %d", (uint16_t)c);
+        cr_assert(std::holds_alternative<cmspk::term::VtInputUnknown>(*vtin), "Failed for input value %d", (uint16_t)c);
         cmspk::term::VtInputUnknown vtu = std::get<cmspk::term::VtInputUnknown>(*vtin);
         char8_t raw = vtu.rawValue;
-        cr_assert((c == raw), "Failed for input value %d", (uint8_t)c);
-        cr_assert(not(dut.canGetData()), "Failed for input value %d", (uint8_t)c);
+        cr_assert((c == raw), "Failed for input value %d", (uint16_t)c);
+        cr_assert(not(dut.canGetData()), "Failed for input value %d", (uint16_t)c);
     }
 }
